pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
local_scheme = "no-local-version"
version_file = "docsearch/_version.py"

[project]
name = "docsearch"
authors = [
    {name = "Logan Lang", email = "lllang@mix.wvu.edu"}
]
description = "DocSearch: An advanced document search and retrieval system leveraging Retrieval-Augmented Generation for intelligent natural language search across PDF document collections."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.8"
keywords = ["rag", "search", "pdf", "llm", "retrieval", "documents", "vector", "embeddings"]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
]

dynamic = ["version"]

dependencies = [
    "python-dotenv",
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "requests",
    "beautifulsoup4",
    "llama-index",
    "openai",
    "PyPDF2",
    "pdf2image",
    "Pillow",
]

[project.optional-dependencies]
dev = [
    "autopep8",
    "jupyterlab",
    "ipywidgets",
    "pytest",
    "black",
    "isort",
    "sphinx",
    "sphinx_rtd_theme",
]

web = [
    "selenium",
    "beautifulsoup4",
]

viz = [
    "plotly",
    "wordcloud",
]

all = [
    "documentrag[dev,web,viz]",
]

[project.urls]
Repository = "https://github.com/loganlang/DocSearch"
Issues = "https://github.com/loganlang/DocSearch/issues"
Documentation = "https://docs.docsearch.ai"

[tool.setuptools.packages.find]
where = ["."]
include = ["docsearch*"]
exclude = ["docsearch.tests*"]

[tool.setuptools.package-data]
docsearch = ["**/*.json", "**/*.csv", "**/*.yml"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88